Add caf::chunk and re-implement caf.net frames
