+++ libcaf_net/caf/net/web_socket/frame.hpp
@@ -6,6 +6,7 @@
 
 #include "caf/async/fwd.hpp"
 #include "caf/byte_span.hpp"
+#include "caf/chunk.hpp"
 #include "caf/config.hpp"
 #include "caf/detail/net_export.hpp"
 #include "caf/fwd.hpp"
@@ -17,92 +18,17 @@
 #include <cstddef>
 #include <string_view>
 
-#ifdef CAF_CLANG
-#  pragma clang diagnostic push
-#  pragma clang diagnostic ignored "-Wc99-extensions"
-#elif defined(CAF_GCC)
-#  pragma GCC diagnostic push
-#  pragma GCC diagnostic ignored "-Wpedantic"
-#elif defined(CAF_MSVC)
-#  pragma warning(push)
-#  pragma warning(disable : 4200)
-#endif
-
 namespace caf::net::web_socket {
 
 /// An implicitly shared type for passing along WebSocket data frames, i.e.,
 /// text or binary frames.
 class CAF_NET_EXPORT frame {
 public:
-  // -- member types -----------------------------------------------------------
+  // -- factory functions ------------------------------------------------------
 
-  class data {
-  public:
-    data() = delete;
-
-    data(const data&) = delete;
-
-    data& operator=(const data&) = delete;
-
-    data(bool is_bin, size_t size) : rc_(1), bin_(is_bin), size_(size) {
-      static_cast<void>(padding_); // Silence unused-private-field warning.
-    }
-
-    explicit data(const_byte_span bytes);
-
-    explicit data(std::string_view str);
-
-    data(size_t total_size, span<const const_byte_span> bufs);
-
-    data(size_t total_size, span<const std::string_view> bufs);
-
-    // -- reference counting ---------------------------------------------------
-
-    bool unique() const noexcept {
-      return rc_ == 1;
-    }
-
-    void ref() const noexcept {
-      rc_.fetch_add(1, std::memory_order_relaxed);
-    }
-
-    void deref() noexcept;
-
-    friend void intrusive_ptr_add_ref(const data* ptr) {
-      ptr->ref();
-    }
-
-    friend void intrusive_ptr_release(data* ptr) {
-      ptr->deref();
-    }
-
-    // -- properties -----------------------------------------------------------
-
-    size_t size() noexcept {
-      return size_;
-    }
-
-    bool is_binary() const noexcept {
-      return bin_;
-    }
-
-    std::byte* storage() noexcept {
-      return storage_;
-    }
-
-    const std::byte* storage() const noexcept {
-      return storage_;
-    }
-
-  private:
-    static constexpr size_t padding_size = CAF_CACHE_LINE_SIZE
-                                           - sizeof(std::atomic<size_t>);
-    mutable std::atomic<size_t> rc_;
-    std::byte padding_[padding_size];
-    bool bin_;
-    size_t size_;
-    std::byte storage_[];
-  };
+  static frame from_chunk(chunk ch) {
+    return frame{std::move(ch).get_data()};
+  }
 
   // -- constructors, destructors, and assignment operators --------------------
 
@@ -116,7 +42,9 @@ public:
 
   frame& operator=(const frame&) = default;
 
-  explicit frame(const_byte_span bytes);
+  explicit frame(const_byte_span bytes) : data_(chunk{bytes}.get_data()) {
+    // nop
+  }
 
   explicit frame(std::string_view text);
 
@@ -166,8 +94,16 @@ public:
 
   std::string_view as_text() const noexcept;
 
+  chunk as_chunk() const& {
+    return chunk{data_};
+  }
+
+  chunk as_chunk() && {
+    return chunk{std::move(data_)};
+  }
+
 private:
-  explicit frame(intrusive_ptr<data> ptr) : data_(std::move(ptr)) {
+  explicit frame(intrusive_ptr<chunk::data> ptr) : data_(std::move(ptr)) {
     // nop
   }
 
@@ -178,15 +114,7 @@ private:
   template <class... Args>
   void init(size_t payload_size, Args&&... arg);
 
-  intrusive_ptr<data> data_;
+  intrusive_ptr<chunk::data> data_;
 };
 
 } // namespace caf::net::web_socket
-
-#ifdef CAF_CLANG
-#  pragma clang diagnostic pop
-#elif defined(CAF_GCC)
-#  pragma GCC diagnostic pop
-#elif defined(CAF_MSVC)
-#  pragma warning(pop)
-#endif
